+++ Packages/BrandonUtils/Standalone/Exceptions/EnumNotInSetException.cs
@@ -5,13 +5,22 @@ using System.Linq;
 
 using BrandonUtils.Standalone.Collections;
 
+using JetBrains.Annotations;
+
 namespace BrandonUtils.Standalone.Exceptions {
-    public class EnumNotInSubsetException<T> : InvalidEnumArgumentException where T : Enum {
+    [Obsolete("Please use " + nameof(EnumNotInSetException<DayOfWeek>) + " instead")]
+    public class EnumNotInSubsetException<T> : EnumNotInSetException<T> where T : Enum {
+        public EnumNotInSubsetException(ICollection<T> subset, IEnumerable<T> values = null, string message = null, Exception innerException = null) : base(subset, values, message, innerException) { }
+        public EnumNotInSubsetException(ICollection<T> subset, T              invalidValue,  string message = null, Exception innerException = null) : base(subset, invalidValue, message, innerException) { }
+    }
+
+    [PublicAPI]
+    public class EnumNotInSetException<T> : InvalidEnumArgumentException where T : Enum {
         public override string Message { get; }
 
         /// <inheritdoc cref="InvalidEnumArgumentException"/>
         /// <summary>
-        /// Constructs a new <see cref="EnumNotInSubsetException{T}"/>, listing information about the invalid <paramref name="values"/> and the <paramref name="subset"/>.
+        /// Constructs a new <see cref="EnumNotInSetException{T}"/>, listing information about the invalid <paramref name="values"/> and the <paramref name="subset"/>.
         /// </summary>
         /// <param name="values">A collection of <see cref="T"/> values, where <b>at least one</b> (but not necessarily all) is <b>not</b> in <paramref name="subset"/>.
         /// <br/>
@@ -20,11 +29,11 @@ namespace BrandonUtils.Standalone.Exceptions {
         /// <param name="subset">The set of valid <see cref="T"/> values.</param>
         /// <param name="message">A user-provided message, which will be <b>prepended</b> to the built-in message.</param>
         /// <param name="innerException">The <see cref="Exception"/> that caused this, if any.</param>
-        public EnumNotInSubsetException(
+        public EnumNotInSetException(
             ICollection<T> subset,
-            IEnumerable<T> values = null,
-            string message = null,
-            Exception innerException = null
+            IEnumerable<T> values         = null,
+            string         message        = null,
+            Exception      innerException = null
         ) : base(message, innerException) {
             var badValues = (values ?? Enumerable.Empty<T>()).Where(value => !subset.Contains(value)).ToArray();
 
@@ -41,11 +50,11 @@ namespace BrandonUtils.Standalone.Exceptions {
             Message = lines.Where(ln => ln != null).JoinString("\n");
         }
 
-        public EnumNotInSubsetException(
+        public EnumNotInSetException(
             ICollection<T> subset,
-            T invalidValue,
-            string message = null,
-            Exception innerException = null
+            T              invalidValue,
+            string         message        = null,
+            Exception      innerException = null
         ) : this(subset, Enumerable.Repeat(invalidValue, 1), message, innerException) { }
     }
 }